Moving Facebook helpers to separate file.
