+++ resources/assets/components/actions/ShareAction/ShareAction.js
@@ -9,7 +9,6 @@ import Embed from '../../utilities/Embed/Embed';
 import Modal from '../../utilities/Modal/Modal';
 import PostCreatedModal from '../PostCreatedModal';
 import { formatPostPayload } from '../../../helpers/forms';
-import { loadFacebookSDK } from '../../../helpers/facebook';
 import PrimaryButton from '../../utilities/Button/PrimaryButton';
 import { SOCIAL_SHARE_TYPE } from '../../../constants/post-types';
 import TextContent from '../../utilities/TextContent/TextContent';
@@ -20,8 +19,11 @@ import {
   trackAnalyticsEvent,
 } from '../../../helpers/analytics';
 import {
-  dynamicString,
+  loadFacebookSDK,
   showFacebookShareDialog,
+} from '../../../helpers/facebook';
+import {
+  dynamicString,
   showTwitterSharePrompt,
   withoutNulls,
 } from '../../../helpers';
