+++ resources/assets/components/utilities/Share/Share.js
@@ -4,10 +4,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import {
-  handleFacebookShareClick,
-  handleTwitterShareClick,
-} from '../../../helpers';
+import { handleTwitterShareClick } from '../../../helpers';
+import { handleFacebookShareClick } from '../../../helpers/facebook';
 
 import './share.scss';
